Day 10 - Handle edge case of ordering the asteroids by distance
This did not actually impact me during the competition as I got the correct answer without implementing this.  But I was aware it was possible I would need to.  I think I just got lucky with my input, only takes a few seconds to write this bit though so not sure how much time it actually saved me.